rename terrascript client to terrascript aws client
