+++ reconcile/terraform_aws_route53.py
@@ -9,7 +9,8 @@ from reconcile.status import ExitCodes
 from reconcile.utils import dnsutils
 from reconcile.utils.semver_helper import make_semver
 from reconcile.utils.defer import defer
-from reconcile.utils.terrascript_client import TerrascriptClient as Terrascript
+from reconcile.utils.terrascript_aws_client \
+    import TerrascriptClient as Terrascript
 from reconcile.utils.terraform_client import TerraformClient as Terraform
