+++ reconcile/terraform_tgw_attachments.py
@@ -9,7 +9,8 @@ from reconcile.utils.aws_api import AWSApi
 from reconcile.utils.defer import defer
 from reconcile.utils.ocm import OCMMap
 from reconcile.utils.terraform_client import TerraformClient as Terraform
-from reconcile.utils.terrascript_client import TerrascriptClient as Terrascript
+from reconcile.utils.terrascript_aws_client \
+    import TerrascriptClient as Terrascript
 from reconcile.utils.semver_helper import make_semver
