+++ reconcile/terraform_users.py
@@ -8,7 +8,8 @@ from reconcile.utils.smtp_client import SmtpClient
 from reconcile import queries
 
 from reconcile.utils.semver_helper import make_semver
-from reconcile.utils.terrascript_client import TerrascriptClient as Terrascript
+from reconcile.utils.terrascript_aws_client \
+    import TerrascriptClient as Terrascript
 from reconcile.utils.terraform_client import TerraformClient as Terraform
 
 TF_POLICY = """
